+++ lib/bobs_bloc.dart
@@ -1,1 +1,3 @@
-export 'src/request_cubit.dart';
+export 'src/cubit.dart';
+export 'src/state.dart';
+export 'src/status.dart';