+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0 - 20.12.25
+
+- Rename `BobsRequestCubitState` to `BobsBlocState`
+- Rename `BobsRequestCubitMixin` to `BobsCubitMixin`
+
 ## 0.0.1 - 14.12.25
 
 - Initial version.